requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
chromadb
pydantic
requests
pypdf